requirements.txt
matplotlib==3.4.2
scipy==1.6.1
tensorflow_gpu==2.4.1
numpy==1.19.5
Keras==2.4.3
Pillow==8.3.1
PyQt5==5.15.4
PyqtWebEngine

#tensorflow==2.5.0
